Fix minor UI issues
