+++ flags.d.ts
@@ -7,5 +7,6 @@ declare module 'launchdarkly-js-sdk-common' {
     bitbucketSync?: boolean;
     tokenFlowButton?: boolean;
     swapStylesAlpha?: boolean;
+    genericVersionedAlpha?: boolean;
   }
 }
