+++ src/app/components/BrokenReferenceIndicator.tsx
@@ -1,22 +1,7 @@
 import React from 'react';
-import { styled } from '@/stitches.config';
-import IconBrokenLink from '@/icons/brokenlink.svg';
 import { SingleToken } from '@/types/tokens';
 import { TokensContext } from '@/context';
-
-const StyledIndicator = styled('div', {
-  position: 'absolute',
-  top: '3px',
-  right: '3px',
-  borderRadius: '100%',
-  border: '1px solid $bgDefault',
-  background: '$dangerFg',
-  width: '6px',
-  height: '6px',
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'center',
-});
+import StyledBrokenReferenceIndicator from './StyledBrokenReferenceIndicator';
 
 type Props = {
   token: SingleToken;
@@ -31,9 +16,7 @@ export default function BrokenReferenceIndicator({ token }: Props) {
 
   if (failedToResolve) {
     return (
-      <StyledIndicator>
-        <IconBrokenLink />
-      </StyledIndicator>
+      <StyledBrokenReferenceIndicator />
     );
   }
   return null;
